requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==20.2.0
Babel==2.11.0
backports.zoneinfo==0.2.1
certifi==2022.9.24
cffi==1.15.1
channels==3.0.5
channels-redis==4.0.0
charset-normalizer==2.1.1
colorama==0.4.6
constantly==15.1.0
cryptography==38.0.1
daphne==3.0.2
defusedxml==0.7.1
Deprecated==1.2.13
Django==3.2.7
django-allauth==0.51.0
django-auto-prefetch==1.3.0
django-crispy-forms==1.14.0
django-debug-toolbar==3.7.0
django-formtools==2.4
django-htmx==1.12.2
django-otp==1.1.4
django-phonenumber-field==6.4.0
django-redis==5.2.0
django-two-factor-auth==1.14.0
django-user-sessions==1.7.1
django-webpush==0.3.4
django-xforwardedfor-middleware==2.0
frozenlist==1.3.3
geoip2==4.6.0
http-ece==1.1.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
maxminddb==2.2.0
msgpack==1.0.4
multidict==6.0.2
oauthlib==3.2.2
packaging==21.3
phonenumbers==8.13.0
Pillow==9.2.0
psycopg2==2.9.4
py-vapid==1.8.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.6.0
pyOpenSSL==22.1.0
pyparsing==3.0.9
python3-openid==3.2.0
pytz==2022.5
pywebpush==1.9.4
qrcode==7.3.1
redis==4.3.4
requests==2.28.1
requests-oauthlib==1.3.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
Twisted==22.8.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing-extensions==4.4.0
tzdata==2022.5
urllib3==1.26.12
wrapt==1.14.1
yarl==1.8.1
zope.interface==5.5.0
